pyproject.toml
[project]
name = "debating-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.0.4",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.6",
    "langgraph>=0.5.0",
    "streamlit>=1.46.1",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
]
